Focus on synchronous programs blocking
This change introduces synchronous programming as a blocking style, more
in line with the contrast drawn in the following paragraph.  The prior
phrasing focused on linear program order execution, presumably to draw a
contrast with the state machine model that a lot of asynchronous
programs follow.  While interesting, the context really is a discussion
of blocking/non-blocking.